requirements.txt
pandas
numpy
scikit-learn
streamlit
python-dotenv
requests